% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_occurence.R
\name{barplotOccurrence.f}
\alias{barplotOccurrence.f}
\title{Fréquences d'occurrence (données agrégées par espèces)}
\usage{
barplotOccurrence.f(
  factGraph,
  factGraphSel,
  listFact,
  listFactSel,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{factGraph}{chr, facteur de séparation des graphiques}

\item{factGraphSel}{chr, sélection de modalité du facteur de séparation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalités sélectionnées pour le(s) facteur(s) de regroupement}

\item{dataEnv}{environnement de stockage des données}

\item{baseEnv}{environnement parent}
}
\value{
none
}
\description{
Cette fonction permet la création du barplot des fréquences d'occurrence.
}
