% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_familles.R
\name{barplotOccurrenceFamille.f}
\alias{barplotOccurrenceFamille.f}
\title{Barplot des frequences de familles}
\usage{
barplotOccurrenceFamille.f(
  factGraph,
  factGraphSel,
  fact,
  factSel,
  families = NA,
  new_window = TRUE,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{factGraphSel}{chr, selection de modalites du facteur de separation des graphiques}

\item{fact}{chr, facteur de regroupement}

\item{factSel}{chr, modalites selectionnees pour le facteur de regroupement}

\item{families}{vector}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}

\item{factGaph}{chr, facteur de separation des graphs}
}
\description{
Creation d'un barplot de frequences d'occurence pour les familles
}
