% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_occurence.R
\name{freq_occurrence.f}
\alias{freq_occurrence.f}
\title{Frequences d'occurrence}
\usage{
freq_occurrence.f(
  agregation,
  factGraph,
  factGraphSel = NA,
  listFact,
  listFactSel = NA,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{agregation}{chr, type d'agregation, par espece (espece) ou par groupe d'espece (unitobs)}

\item{factGraph}{chr, facteur de separation des graphique.}

\item{factGraphSel}{chr, selection de modalites du facteur de separation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalite selectionnees pour le(s) facteur(s) de regroupement}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Creation d'un barplot de frequence d'occurrence selon les parametres fournis
}
\details{
Fonction pour tracer un barplot de frequences d'occurrence pour des donnees agregees par especes.

agregation peut prendre la valeur 'espece' ou 'unitobs'. Si 'espece', l'agregation se fera
par especes, si unitobs, elle se fera par groupe d'especes.

factGraphSel peut prendre differentes valeurs en fonction de celle de factGraph. Ce parametre
est facultatif. S'il est egal a NA, toutes les modalites du facteur seront selectionnees.

listFact ne peut pas avoir plus de 2 facteurs.

listFactSel peut prendre differentes valeurs en fonction de celle de listFact. Ce parametre est
facultatif. S'il est egal a NA, alors toutes les modalites du ou des facteur(s) de regroupement
selectionne(s) seront prises en compte.
}
\examples{
freq_occurrences.f(agregation = "espece", factGraph = c("scient.name"),
  factGraphSel = c("Chromis_chromis"), listFact = c("year", "protection.status"),
  listFactSel = list(c("2011", "2019"), NA), dataEnv = .dataEnv, baseEnv = .baseEnv)

}
