% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_modele.R
\name{lm.f}
\alias{lm.f}
\title{Modele lineaire (donnees agregees par especes)}
\usage{
lm.f(
  agregation,
  metrique,
  factAna,
  factAnaSel = NA,
  listFact,
  listFactSel = NA,
  tableMetrique,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{agregation}{chr, type d'agregation, par espece (espece) ou par groupe d'espece (unitobs)}

\item{metrique}{chr, metrique choisie}

\item{factAna}{chr, facteur de separation des graphiques}

\item{factAnaSel}{chr, selection de modalites du facteur de separation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalite(s) selectionnee(s) pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de metrique}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Calcul de modeles lineaires pour des donnees agregees par especes.
}
\details{
Fonction pour calculer des modeles lineaires pour des donnees agregees par especes.

agregation peut prendre la valeur 'espece' ou 'unitobs'. Si 'espece', l'agregation se fera
par especes, si unitobs, elle se fera par groupe d'especes.

tableMetrique peut prendre les valeurs 'unitSp' (/station /especes /classe de taille),
'unitSpSz' (/station /especes) ou 'unit' (de biodiversite (/station)). Ce dernier cas n'est
possible uniquement lorsque agregation == 'unitobs'.

metrique peut prendre differentes valeur en fonction de celle de tableMetrique. Elle peut prendre
les valeurs 'density', 'max.density', 'max.number', 'number', 'sd.density', 'sd.number' lorsque
tableMetrique == 'unitSp'. Lorsque tableMetrique == 'unitSpSz', metrique peut prendre ces memes
valeurs ainsi que la valeur 'abundance.propo.SC'. Lorsque tableMetrique == 'unit', metrique peut
prendre les valeurs suivantes : 'species.richness', 'simpson', 'simpson.l', 'pielou', 'hill',
'Delta', 'DeltaStar', 'LambdaPlus', 'DeltaPlus', 'SDeltaPlus', 'relative.SR.site',
'relative.SR.site.phylum', 'relative.SR.data', 'relative.SR.region', 'relative.SR.region.phylum'.

factAnaSel peut prendre differentes valeurs en fonction de celle de factAna. Ce parametre
est facultatif. S'il est egal a NA, toutes les modalites du facteur seront selectionnees.

Il est deconseille de mettre plus de deux facteurs pour listFact, et il n'est pas possible d'en
mettre plus de trois.

listFactSel peut prendre differentes valeurs en fonction de celle de listFact. Ce parametre est
facultatif. S'il est egal a NA, alors toutes les modalites du ou des facteur(s) de regroupement
selectionne(s) seront prises en compte.
}
\examples{
lm.f(agregation = "espece", metrique = c("density"), factAna = c("scient.name"),
  factAnaSel = c("Chromis_chromis"), listFact = c("year", "protection.status"),
  listFactSel = list(c("2011", "2019"), NA), tableMetrique = "unitSp",
  dataEnv = .dataEnv, baseEnv = .baseEnv)

}
