% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_files.R
\name{load_files.f}
\alias{load_files.f}
\title{Importation des donnees et calculs des poids}
\usage{
load_files.f(filePathes, dminMax = 5, dataEnv, baseEnv)
}
\arguments{
\item{filePathes}{chr, vecteur nomme contenant les chemins des fichiers :
\itemize{unitobs}{ : chr, fichier unitobs}
\itemize{obs}{ : chr, fichier d'observation}
\itemize{refesp}{ : chr, referentiel especes}
\itemize{refspa}{ : chr, referentiel spatial (facultatif)}
\itemize{ws}{ : chr, dossier de travail}
\itemize{...}}

\item{dminMax}{seuil de Min.Distance (en m) au-dessus duquel les observations ne sont pas prises en compte}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\value{
list, liste nommee contenant les tables de donnees :
\itemize{obs}{ : donnees d'observation}
\itemize{unitobs}{ : unites d'observation}
\itemize{refesp}{ : referentiel espece}
}
\description{
Chargement des donnees depuis les chemins des fichiers et calcul des poids.
}
\details{
Cette fonction permet de faire le chargement des donnees.

Elle prend en entree les chemins des differents fichiers (sous forme d'un
vecteur nomme).
Elle prend egalement en entree deux environnements, celui où stocker les donnees, et
l'environnement parent. (Pas forcement tres pratique d'utilisation, a enlever plus tard ?)

Dans un premier temps, elle va charger les differents fichiers de donnees (obs, unitobs, refesp,
refspa). Elle va ensuite faire le calcul des tables de metriques. Puis, elle va enregistrer les
tableaux de donnees dans l'environnement dataEnv afin de pouvoir les reutiliser plus tard.
Enfin, elle va faire le calcul des poids pour chaque AMP.

Elle renvoie ensuite les differentes tables de donnees, sous forme d'une liste nommee.
}
\examples{
ws_path <- system.file('example_data/COTE BLEUE/Data/', package = 'PAMPA')
unitobs_path <- system.file('example_data/COTE BLEUE/Data/UnitObs_Staviro_CB191110_280720.txt', package = 'PAMPA')
obs_path <- system.file('example_data/COTE BLEUE/Data/Obs_Staviro_CB191110_AllIdentifiedSpecies_280720.txt', package = 'PAMPA')
refesp_path <- system.file('example_data/COTE BLEUE/Data/refEspecesMed_general_270720 - Int peche est intCH_CB.txt', package = 'PAMPA')

filepaths <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path,
  refspa = NULL, ws = ws_path)

load_files.f(filepaths, dminMax = 5, .dataEnv, .baseEnv)

}
