% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilang_documentation.R
\name{load_files.f_en}
\alias{load_files.f_en}
\title{Data importation and weight calculation}
\usage{
load_files.f_en()
}
\arguments{
\item{filePathes}{chr, named vector whith the file path :
\itemize{unitobs}{ : chr, unitobs file}
\itemize{obs}{ : chr, observation file}
\itemize{refesp}{ : chr, species reference table}
\itemize{refspa}{ : chr, spatial reference table (optional)}
\itemize{ws}{ : chr, working directory}
\itemize{...}}

\item{dminMax}{minimal threshold of Min.Distance (in m) above which observations are not tacking in consideration}

\item{dataEnv}{storage data environment
environnement de stockage des donnees}

\item{baseEnv}{main environment}
}
\value{
list, nammed list with the folowing data frame :
\itemize{obs}{ : observation data}
\itemize{unitobs}{ : units observation}
\itemize{refesp}{ : species reference table}
}
\description{
Load data from the file paths and calculation of weight
}
\examples{
ws_path <- system.file('example_data/COTE BLEUE/Data/', package = 'PAMPA')
unitobs_path <- system.file('example_data/COTE BLEUE/Data/UnitObs_Staviro_CB191110_280720.txt', package = 'PAMPA')
obs_path <- system.file('example_data/COTE BLEUE/Data/Obs_Staviro_CB191110_AllIdentifiedSpecies_280720.txt', package = 'PAMPA')
refesp_path <- system.file('example_data/COTE BLEUE/Data/refEspecesMed_general_270720 - Int peche est intCH_CB.txt', package = 'PAMPA')

filepaths <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path,
  refspa = NULL, ws = ws_path)

load_files.f(filepaths, dminMax = 5, .dataEnv, .baseEnv)
}
