% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multilang_documentation.R
\name{load_files.f_fr}
\alias{load_files.f_fr}
\title{Importation des donnees et calculs des poids}
\usage{
load_files.f_fr()
}
\arguments{
\item{filePathes}{chr, vecteur nomme contenant les chemins des fichiers :
\itemize{unitobs}{ : chr, fichier unitobs}
\itemize{obs}{ : chr, fichier d'observation}
\itemize{refesp}{ : chr, referentiel especes}
\itemize{refspa}{ : chr, referentiel spatial (facultatif)}
\itemize{ws}{ : chr, dossier de travail}
\itemize{...}}

\item{dminMax}{seuil de Min.Distance (en m) au-dessus duquel les observations ne sont pas prises en compte}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\value{
list, liste nommee contenant les tables de donnees :
\itemize{obs}{ : donnees d'observation}
\itemize{unitobs}{ : unites d'observation}
\itemize{refesp}{ : referentiel espece}
}
\description{
Chargement des donnees depuis les chemins des fichiers et calcul des poids.
}
\examples{
ws_path <- system.file('example_data/COTE BLEUE/Data/', package = 'PAMPA')
unitobs_path <- system.file('example_data/COTE BLEUE/Data/UnitObs_Staviro_CB191110_280720.txt', package = 'PAMPA')
obs_path <- system.file('example_data/COTE BLEUE/Data/Obs_Staviro_CB191110_AllIdentifiedSpecies_280720.txt', package = 'PAMPA')
refesp_path <- system.file('example_data/COTE BLEUE/Data/refEspecesMed_general_270720 - Int peche est intCH_CB.txt', package = 'PAMPA')

filepaths <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path,
  refspa = NULL, ws = ws_path)

load_files.f(filepaths, dminMax = 5, .dataEnv, .baseEnv)
}
