% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{maps.f}
\alias{maps.f}
\title{Cartes}
\usage{
maps.f(
  agregation,
  graphType,
  metrique,
  factSpatial,
  factSpatialSel = NA,
  factGraph = NULL,
  factGraphSel = NA,
  listFact,
  listFactSel = NA,
  tableMetrique,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{agregation}{chr, type d'agregation par espece (espece) ou par groupe d'espece (unitobs)}

\item{graphType}{chr, type de graphique (boxplot, barplot, symboles, couleurs)}

\item{metrique}{chr, metrique choisie}

\item{factSpatial}{chr, facteur de regroupement spatial}

\item{factSpatialSel}{chr, selection pour le facteur de regroupement spatial}

\item{factGraph}{chr, facteur de selection des especes}

\item{factGraphSel}{chr, selection de modalites pour le facteur de selection des especes}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{chr, modalites selectionnees pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de metriques}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}

\item{coord}{num, emprise spatiale (bounding box). Si NULL, toute la carte. Vecteur nomme :
\itemize{
  \item{W}{ : longitude minimale (coordonnee ouest) en degre. Negatif pour l'ouest de Greenwich.}
  \item{E}{ : longitude maximale (coordonnee est) en degre. Negatif pour l'ouest de Greenwich.}
  \item{N}{ : latitude maximale (coordonnee nord) en degre. Negatif pour l'hemisphere sud'}
  \item{S}{ : latitude minimale (coordonnee sud) en degre. Negatif pour l'hemisphere sud'}
}}
}
\description{
Fonction permettant de creer des cartes en fonctions des parametres donnes
par l'utilisateur
}
\details{
Fonction pour tracer des cartes.

\code{agregation} peut prendre la valeur 'espece' ou 'unitobs'. Si 'espece', l'agregation se fera
par especes, si 'unitobs', elle se fera par groupe d'especes.

\code{graphType} peut prendre la valeur 'boxplot', 'barplot', 'symboles' ou 'couleurs'.
Si 'boxplot' ou 'barplot', le(s) boxplot(s) ou barplot(s) seront affiches sur la carte.
Si symboles, la metrique choisie sera representee par des cercles, si couleurs, elle sera
representee par des couleurs.

\code{tableMetrique} peut prendre les valeurs 'unitSp' (/station /especes /classe de taille),
'unitSpSz' (/station /especes) ou 'unit' (de biodiversite (/station)). Ce dernier cas n'est
possible uniquement lorsque \code{agregation == 'unitobs'}. La table de metriques 'unitSpSz'
n'est pas disponible si le jeu de donnees est un jeu de donnees de benthos.

\code{metrique} peut prendre differentes valeurs en fonction de celle de tableMetrique et
du jeu de donnees.

\code{factSpatialSel} peut prendre differentes valeurs en fonction de celle de factSpatial.
Ce parametre est facultatif. S'il est egal a NA, toutes les modalites du facteur seront
selectionnees.

\code{factGraph} est facultatif.

\code{factGraphSel} peut prendre differentes valeurs en fonction de celle de factGraph. Ce
parametre est facultatif. S'il est egal a NA, toutes les modalites du facteur seront
selectionnees.

\code{listFact} n'est pas necessaire dans le cas ou \code{graphType} est egal a 'symboles' ou
'couleurs'. De meme pour \code{listFactSel}.

\code{listFactSel} peut prendre differentes valeurs en fonction de celle de \code{listFact}. Ce
parametre est facultatif. S'il est egal a NA, alors toutes les modalites du ou des facteur(s) de
regroupement selectionne(s) seront prises en compte.
}
\examples{
maps.f(agregation = "espece", graphType = "barplot", metrique = "density",
  factSpatial = "CODE.SITE", factSpatialSel = c("GN_07_02", "GN_07_04", "GN_07_05", "GN_07_06"),
  factGraph = "family", factGraphSel = "Cheloniidae", listFact = "year", listFactSel = NA,
  tableMetrique = "unitSp", coord = c(W = 166.15, E = 166.55, S = -22.40, N = -22.15),
  .dataEnv, .baseEnv)

}
