% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_modele.R
\name{modeleLineaireWP2.esp.f}
\alias{modeleLineaireWP2.esp.f}
\title{Modèle linéaire (données agrégées par espèces)}
\usage{
modeleLineaireWP2.esp.f(
  metrique,
  factAna,
  factAnaSel,
  listFact,
  listFactSel,
  tableMetrique,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{chr, métrique choisie}

\item{factAna}{chr, facteur de séparation des graphiques}

\item{factAnaSel}{chr, sélection de modalités pour le factuer de sélection des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalités sélectionnées pour le(s) facteur(s) de regroupement}

\item{dataEnv}{environnement de stockage des donnnées}

\item{baseEnv}{environnement parent}
}
\value{
none
}
\description{
Cette fonction permet de calculer des modèles linéaires sur les données agrégées
par espèces
}
