% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_modele.R
\name{modeleLineaireWP2.unitobs.f}
\alias{modeleLineaireWP2.unitobs.f}
\title{Modèle linéaire (données agrégées par unités d'observation)}
\usage{
modeleLineaireWP2.unitobs.f(
  metrique,
  factAna,
  factAnaSel,
  listFact,
  listFactSel,
  tableMetrique,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{chr, métrique choisie}

\item{factAna}{chr, facteur de séparation des graphiques}

\item{factAnaSel}{chr, sélection de modalités pour le facteur de séparation des graphiques}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{chr, modalités sélectionnées pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de métrique}

\item{dataEnv}{environnement de stockage des données}

\item{baseEnv}{environnement parent}
}
\value{
none
}
\description{
Cette fonction permet de faire des modèles linéaires sur des données agrégées
par unités d'observation
}
