% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_descriptives.R
\name{moyenne_annee_visite.f}
\alias{moyenne_annee_visite.f}
\title{Nombre moyen d'annee de visite pour chaque site}
\usage{
moyenne_annee_visite.f(site = NA, year = NA, dataEnv, baseEnv = .GlobalEnv)
}
\arguments{
\item{site}{chr, selection des sites a etudier, si NA, tous les sites seront selectionnes}

\item{year}{chr, selection des annees a etudier, si NA toutes les annees seront selectionnees}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Cree un graph (et une carte si refspa disponible) presentant le nombre de moyen
d'annee de visite par site
}
\examples{
PAMPA::moyenne_annee_visite.f(site = NA, year = NA, .dataEnv, .baseEnv)

}
