% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permanova.R
\name{permanova_pampa.f}
\alias{permanova_pampa.f}
\title{Permanova}
\usage{
permanova_pampa.f(
  tableMetrique,
  etude,
  fact,
  method = "bray",
  nb_perm = 1000,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{tableMetrique}{chr, le nom de la table de metrique}

\item{etude}{chr, donnees etudiees (presences/absence (pres.abs) ou abondance ("abundance.prop.SC"))}

\item{fact}{chr, liste des facteurs explicatifs}

\item{method}{chr, methode utilisee pour la creation de la matrice de dissimilarites}

\item{nb_perm}{int, nombre de permutations}

\item{dataEnv}{environnemant de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Calcul d'une permanova selon les parametres fournis
}
\details{
Cette fonction permet de faire une permanova.
}
\examples{
PAMPA::permanova_pampa.f("unitSp", "pres.abs", c("site", "year"), method = "bray",
  nb_perm = 1000, .dataEnv, .baseEnv)

}
