% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_data.R
\name{restoreData.f}
\alias{restoreData.f}
\title{Selection et recalcul}
\usage{
restoreData.f(baseEnv, dataEnv)
}
\arguments{
\item{baseEnv}{: environnement parent}

\item{dataEnv}{: environnement de stockage des donnees}
}
\value{
none
}
\description{
Permet la selection des donnees d'observation, selon un critere du
referentiel d'unites d'observation
}
\details{
Fonction pour la restauration des donnees apres la selection et le recalcul.
Vraiment utile ? Un loadData.f suffit pour ca...
Et si on enlève les environnement, sera tocké dans une variable...
}
\examples{
restoreData.f(.baseEnv, .dataEnv)

}
