% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection.R
\name{selection.f}
\alias{selection.f}
\title{Selection des donnees}
\usage{
selection.f(field, baseEnv, dataEnv)
}
\arguments{
\item{field}{: chr, unitobs ou refesp, type de champs choisi pour la selection.}

\item{baseEnv}{: environnement parent}

\item{dataEnv}{: environnement de stockage des donnees}
}
\value{
list, liste contenant les tables de donnees selon la selection.
 \itemize{unitobs}{ : chr, fichier unitobs}
 \itemize{obs}{ : chr, fichier d'observation}
 \itemize{refesp}{ : chr, referentiel especes}
}
\description{
Selection des donnees selon un champs du referentiel espece ou un
champs des unitobs.
}
\details{
Cette fonction permet de faire la selection et le recalcul des donnees.

Elle prend en entree le type de critere selon lequel le recalcul sera fait
(champs du referentiel espece ou champs des unitobs).

Elle prend egalemnt en entree deux environnements, celui ou stocker les donnees,
et l'environnement parent.

Elle va ensuite faire la selection des donnees, en demandant a l'utilisateur
selon quel(s) critere(s) du champs choisi il veut la faire.

Elle ne retourne rien, les donnees sont stockees dans l'environnement .dataEnv
(pas forcement tres pratique, a changer plus tard ?)

A changer plus tard pour supprimer les lignes de code en doublon dans selectionOnRefesp.f (et
dependaces) et selectionOnUnitobs.f (et dependances) ?
}
\examples{
ws_path <- system.file('example_data/COTE BLEUE/Data/', package = 'PAMPA')
unitobs_path <- system.file('example_data/COTE BLEUE/Data/UnitObs_Staviro_CB191110_280720.txt', package = 'PAMPA')
obs_path <- system.file('example_data/COTE BLEUE/Data/Obs_Staviro_CB191110_AllIdentifiedSpecies_280720.txt', package = 'PAMPA')
refesp_path <- system.file('example_data/COTE BLEUE/Data/refEspecesMed_general_270720 - Int peche est intCH_CB.txt', package = 'PAMPA')

filePathes <- c(unitobs = unitobs_path, obs = obs_path, refesp =  refesp_path, refspa = NULL, ws = ws_path)

data <- load_and_calc.f(filePathes, .dataEnv, .baseEnv)

selection.f('unitobs', .baseEnv, .dataEnv)

}
