% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_descriptives.R
\name{station_year.f}
\alias{station_year.f}
\title{Nombre de site visites par an}
\usage{
station_year.f(site = NA, year = NA, dataEnv, baseEnv)
}
\arguments{
\item{site}{chr, selection des sites a etudier, si NA, tous les sites seront selectionnes}

\item{year}{chr, selection des annees a etudier, si NA toutes les annees seront selectionnees}

\item{dataEnv}{environnement de stockage des donnees}

\item{baseEnv}{environnement parent}
}
\description{
Cree un graph (et une carte si refspa disponible) presentant le nombre de visites
de chaque sites par annees
}
\examples{
PAMPA::station_year.f(site = c("Mbe Kouen", "Laregnere"), year = c("2008", "2009", "2010"),
  .dataEnv, .baseEnv)

}
