% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{WP2barplot.unitobs.f}
\alias{WP2barplot.unitobs.f}
\title{Barplot (données agrégées par unités d'observation)}
\usage{
WP2barplot.unitobs.f(
  metrique,
  factGraph,
  factGraphSel,
  listFact,
  listFactSel,
  tableMetrique,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{chr, métrique choisie}

\item{factGraph}{chr, facteur de sélection des espèces}

\item{factGraphSel}{chr, sélection de modalités pour le facteur de sélection des espèces}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalités sélectionnées pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de métriques}

\item{dataEnv}{environnement de stockage des données}

\item{baseEnv}{environnement parent}
}
\description{
Cette fonction permet la création de barplot pour les données agrégées par unités
d'observation.
}
