% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot.R
\name{WP2boxplot.f}
\alias{WP2boxplot.f}
\title{Boxplot (données agrégées par espèces)}
\usage{
WP2boxplot.f(
  metrique,
  factGraph,
  factGraphSel,
  listFact,
  listFactSel,
  tableMetrique,
  dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{: chr, métrique choisie}

\item{factGraph}{: chr, facteur de séparation des graphiques}

\item{factGraphSel}{: chr, sélection de modalité du facteur de séparation des graphiques}

\item{listFact}{: chr, facteur(s) de regroupement}

\item{listFactSel}{: list, modalités sélectionnées pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{: chr,  nom de la table de métrique}

\item{dataEnv}{: environnement de stockage des données}

\item{baseEnv}{: environnement parent}
}
\value{
none
}
\description{
Cette fonction permet la création du boxplot pour les données agrégées par espèces
}
