% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrt.R
\name{WP2MRT.unitobs.f}
\alias{WP2MRT.unitobs.f}
\title{Arbre de régression multivariée (données agrégées par unités d'observation)}
\usage{
WP2MRT.unitobs.f(
  metrique,
  factGraph,
  factGraphSel,
  listFact,
  listFactSel,
  tableMetrique,
  dataEnv = dataEnv,
  baseEnv = .GlobalEnv
)
}
\arguments{
\item{metrique}{chr, métrique choisie}

\item{factGraph}{chr, facteur de sélection des espèces}

\item{factGraphSel}{chr, sélection de modalités pour le facteur de sélection des espèces}

\item{listFact}{chr, facteur(s) de regroupement}

\item{listFactSel}{list, modalité sélectionnées pour le(s) facteur(s) de regroupement}

\item{tableMetrique}{chr, nom de la table de métriques}

\item{dataEnv}{environnement de stockage des données}

\item{baseEnv}{environnement parent}
}
\value{
none
}
\description{
Cette fonction permet de faire des arbres de régression multivariée pour des données
agrégées par unités d'observation
}
